Add the DebugInfo writer
Initial test of the DebugInfo writer.

This writes each "step" to a single zip archive. The debug output may be
large, but is highly compressible so it's beneficial to write it out in
a compressed format. Each item written to the archive get prepended with
a monotonic "step" number, making it simple to order the files.

This also has a StepWriter method, which returns a WriteCloser
that buffers writes and commits them to the archive on Close. This
allows parallel operations to write logs that aren't interleaved with
other operations.

All operations on the DebugInfo  will be safe for concurrent access. All
DebugInfo and related structure methods will also be safe to call on a
nil value, so no additional checks are required throughout the code to
disable the debug info -- leaving the value as nil disables all output.